requirements.txt
streamlit
langchain
sentence-transformers
transformers
torch
faiss-cpu
pypdf
python-docx
tqdm